@@ -11,7 +11,7 @@ authors = [
 maintainers = [
     { name = "Tom Aarsen", email = "tom.aarsen@huggingface.co" }
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 keywords = [
     "Transformer Networks",
     "BERT",
@@ -25,7 +25,6 @@ classifiers = [
     "Development Status :: 5 - Production/Stable",
     "Intended Audience :: Science/Research",
     "License :: OSI Approved :: Apache Software License",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
