@@ -3,7 +3,8 @@ from __future__ import annotations
 import csv
 import logging
 import os
-from typing import TYPE_CHECKING, Callable
+from collections.abc import Callable
+from typing import TYPE_CHECKING
 
 import numpy as np
 import torch
