@@ -1,7 +1,7 @@
 from __future__ import annotations
 
+from collections.abc import Callable
 from enum import Enum
-from typing import Callable
 
 from numpy import ndarray
 from torch import Tensor
