@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 import os
+import tempfile
 from copy import deepcopy
 
 import numpy as np
@@ -10,7 +11,6 @@ from tokenizers import Tokenizer
 from sentence_transformers import SentenceTransformer
 from sentence_transformers.models import Pooling, StaticEmbedding, Transformer
 from sentence_transformers.util import is_datasets_available
-from tests.utils import SafeTemporaryDirectory
 
 if is_datasets_available():
     from datasets import DatasetDict, load_dataset
@@ -108,7 +108,7 @@ def cache_dir():
     if os.environ.get("CI", None):
         # Note: `ignore_cleanup_errors=True` is used to avoid NotADirectoryError in Windows on GitHub Actions.
         # See https://github.com/python/cpython/issues/107408, https://www.scivision.dev/python-tempfile-permission-error-windows/
-        with SafeTemporaryDirectory() as tmp_dir:
+        with tempfile.TemporaryDirectory(ignore_cleanup_errors=True) as tmp_dir:
             yield tmp_dir
     else:
         yield None
