@@ -41,16 +41,16 @@ def test_load_and_encode(model_name: str) -> None:
 
     # Check a known property: encoding a single sentence
     single_sentence_emb = model.encode(["A single sentence."], convert_to_tensor=False)
-    assert isinstance(
-        single_sentence_emb, list
-    ), "Encoding a single sentence with convert_to_tensor=False should return a list of len 1"
+    assert isinstance(single_sentence_emb, list), (
+        "Encoding a single sentence with convert_to_tensor=False should return a list of len 1"
+    )
     assert len(single_sentence_emb) == 1, "Single sentence embedding dict should not be empty"
 
     # If we're using a string instead of a list, we should get a single tensor embedding
     single_sentence_emb_tensor = model.encode("A single sentence.", convert_to_tensor=False)
-    assert isinstance(
-        single_sentence_emb_tensor, Tensor
-    ), "Encoding a single sentence with convert_to_tensor=False should return a tensor"
+    assert isinstance(single_sentence_emb_tensor, Tensor), (
+        "Encoding a single sentence with convert_to_tensor=False should return a tensor"
+    )
     assert single_sentence_emb_tensor.dim() == 1, "Single sentence embedding tensor should be 1D"
 
     # Check encoding with show_progress_bar
